+++ back/src/routes/index.js
@@ -3,9 +3,11 @@ const userRoute = require("./userRoute");
 const canchaRoute = require("./canchaRoutes");
 const reservaRoute = require("./reservaRoute");
 const reviewsRoute = require("./reviewRoute");
+const authRoute = require("./authRoute");
 
 const router = Router();
 
+router.use("/", authRoute);
 router.use("/", userRoute);
 router.use("/canchas", canchaRoute);
 router.use("/reserva", reservaRoute);
