+++ back/src/routes/userRoute.js
@@ -1,7 +1,5 @@
 const { Router } = require("express");
 const bcryptjs = require("bcryptjs");
-const passport = require("passport");
-const jwt = require("jsonwebtoken");
 const {
   getUsersDb,
   deleteUser,
@@ -14,44 +12,6 @@ const { User, Reserva, Cancha } = require("../db");
 const { authMiddleware, adminMiddleware } = require("../middlewares/auth");
 
 const router = Router();
-
-//login local
-router.post("/login", (req, res, next) => {
-  passport.authenticate("local", { session: false }, (err, user, info) => {
-    if (err) {
-      return res.status(500).json({ message: "Error al iniciar sesión" });
-    }
-    if (!user) {
-      return res.status(401).json({ message: info.message });
-    }
-    req.logIn(user, { session: false }, (err) => {
-      if (err) {
-        return res.status(500).json({ message: "Error al iniciar sesión" });
-      }
-      const token = jwt.sign(
-        { userId: user.id, email: user.email },
-        process.env.JWT_SECRET,
-        { expiresIn: "1h" }
-      );
-      return res.status(200).json({ token });
-    });
-  })(req, res, next);
-});
-
-//login Google
-router.get(
-  "/google",
-  passport.authenticate("google", {
-    scope: ["profile", "email"],
-  })
-);
-router.get(
-  "/auth/google/callback",
-  passport.authenticate("google", { failureRedirect: "/" }),
-  (req, res) => {
-    return res.redirect("http://localhost:5173/home");
-  }
-);
 
 //traer users activos-inactivos
 router.get("/users", authMiddleware, getUsersActive);
@@ -98,19 +58,6 @@ router.get("/me", authMiddleware, async (req, res) => {
   }
 });
 
-//cerrar sesión Google
-router.get("/logout", (req, res) => {
-  try {
-    req.logout();
-    req.session.destroy();
-    res.clearCookie("connect.sid", { path: "/" });
-    res.status(200).json({ message: "Haz cerrado sesión con éxito" });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ message: "Error al cerrar sesión" });
-  }
-});
-
 //crear user
 router.post("/users", authMiddleware, async (req, res) => {
   let { name, lastName, email, isAdmin, password, phone } = req.body;
@@ -136,7 +83,7 @@ router.post("/users", authMiddleware, async (req, res) => {
 router.put("/users/:id", authMiddleware, updateUser);
 
 //modificar user autenticado
-router.put("/me", async (req, res) => {
+router.put("/me", authMiddleware, async (req, res) => {
   try {
     const user = await User.findByPk(req.user.id);
     if (!user) {
